Pytest

+++ tests/test_binary.py
@@ -5,6 +5,7 @@ import re
 from subprocess import Popen, PIPE
 from distutils.spawn import find_executable
 
+
 class PstTestCase(unittest.TestCase):
     """This class represents the pst test case"""
 
@@ -12,19 +13,11 @@ class PstTestCase(unittest.TestCase):
     def tearDownClass(cls):
         files = ["trees-root.txt", "trees-pid.txt", "trees-pst.txt"]
         for f in files:
-            try:
-                os.remove(f)
-            except BaseException:
-                print("Error while deleting file ", f)
-
-    def test_file_exists(self):
-        self.assertTrue(find_executable("pst"))
-
-    def test_file_access(self):
-        self.assertEqual(os.access(find_executable("pst"), os.X_OK), True)
-
-    def test_compare_file_size(self):
-        self.assertEqual(os.path.getsize(find_executable("pst")), os.path.getsize('bin/pst'))
+            if os.path.isfile(f):
+                try:
+                    os.remove(f)
+                except BaseException:
+                    print("Error while deleting file ", f)
 
     def test_help_string(self):
         proc = Popen(["pst", "-h"], stdout=PIPE, stderr=PIPE)
@@ -41,7 +34,7 @@ class PstTestCase(unittest.TestCase):
             self.fail("Failed with %s" % error)
         self.assertTrue(os.path.isfile("trees-root.txt"))
         self.assertTrue(os.path.getsize("trees-root.txt") > 0)
-    
+
     def test_compare_output_file_sizes(self):
         proc = Popen(["pst", "-p", "1", "-o", "trees-pid.txt"], stdout=PIPE, stderr=PIPE)
         output, error = proc.communicate()
@@ -51,7 +44,8 @@ class PstTestCase(unittest.TestCase):
         output, error = proc.communicate()
         if error:
             self.fail("Failed with %s" % error)
-        self.assertTrue(os.path.getsize('trees-pst.txt') > os.path.getsize('trees-pid.txt')) 
+        self.assertTrue(os.path.getsize('trees-pst.txt') > os.path.getsize('trees-pid.txt'))
+
 
 if __name__ == "__main__":
-    unittest.main()
+    unittest.main(failfast=True)
